msg/async/rdma: Introduce Device.{cc,h}
Later on, multidevice support will be added and Device class will be the
main RDMA class (instead of Infiniband as it is today).

Issue: 995322
